Added a bunch of legacy table support classes, rewritten from scratch
Not fully tested yet, but these are ancient abandoned table support modules for Swing that got replaced by JavaFX in 2015. These are all rewritten from scratch but have comments on further research, once all of the code is invoked in some GUI wrapper and can be tested properly. Going from 0% to 100% coverage with Javadocs of course is what took most of the time and effort.