Continuing to add legacy modules to the GUI toolkit
Bringing lots more legacy code up-to-date, but not able to take the time to thoroughly test and to verify commented-out code and other questions that came up, as this code hasn't been actively used in at least five years by now due to the JavaFX conversion being completed 100% by 2015. Some of this code may be redundant based on the continued evolution of Java since version 1.3 or 1.4, which is around the timeframe that much of this code was first written. It is provided in case it becomes needed as I pull in yet more legacy code to try to flesh out this library beyond the shell of what I was still using from AWT and Swing these past few years. Doing so also provides more context and scope for making better overall API decisions and library structure.