Continuing to add more legacy files for table support
As before, these are legacy Swing files that have been reformatted and rewritten as well as heavily documented and additionaly have been made to adhere to the style of this library and to derive from existing class hierarchies where possible. There are only a few more table-related classes to bring up to date so that they work in this library.