Final edits to get Swing and JavaFX versions in sync
Although not released yet, the JavaFX version of the table handling is now completed in terms of being brought up to date with these new coding and documentation standards, so then the Swing version was brought into conformance with that version as it has several more years of real-world development and problem solving behind it. The differences between JTable and TableView are minimal in comparison.