Tried to get all archives menu, button, and tool bar boilerplate code posted ASAP
I need this in a work project, even if just as a quick reference on some isolated best practices here and there, but most of the code dates from 2006 to 2008 and may be naive compared to newer contributions and especially JavaFX alternatives. Nevertheless, it is better to get it all posted first, and then clean up or delete after deeper review. In many cases, I defer to Apache libraries once I find a good solution there.