Update TabUtilities.java
Apply lessons learned from first integration pass in a real app. This revised fix appears to take care of issues of tabs not being reachable by component comparison searches. Part of the issue with JTabbedPane is confusing terminology, as the so-called "TabComponent" is actually the little tab that appears in the tab bar, as opposed to the content pane, which is the Tab!